[#3] refactor : service 구조 변경
